Modificações para os gráficos do paper

+++ main_para_programadores.py
@@ -528,9 +528,9 @@ df1 = pd.DataFrame(data=d)
 df1.to_excel("output_transit_depth.xlsx")
 
 # imprimindo valores da profundidade de trânsito em txt
-np.savetxt('output_transit_depth(trans_lat' + str(int(latsugerida)) + 'graus,f_spot=' + str(f_spot * 100) + '%).txt', D_lambda, delimiter=',')
+np.savetxt('output_transit_depth(trans_lat' + str(int(latsugerida)) + 'graus,f_spot=' + str(int(f_spot * 100)) + '%).txt', D_lambda, delimiter=',')
 # imprimindo valores de epsilon de Rackham em txt
-np.savetxt('output_epsilon_Rackham(trans_lat' + str(int(latsugerida)) + 'graus,f_spot=' + str(f_spot * 100) + '%).txt', epsilon_Rackham, delimiter=',')
+np.savetxt('output_epsilon_Rackham(trans_lat' + str(int(latsugerida)) + 'graus,f_spot=' + str(int(f_spot * 100)) + '%).txt', epsilon_Rackham, delimiter=',')
 # imprimindo valores dos comprimentos de onda em txt (útil para construção dos gráficos)
 np.savetxt('output_wavelengths.txt', lambdaEff_nm, delimiter=',')
